fixing toNumber overflow + faster toString implementation
toNumber() returned invalid (negative) value if it reached 32th bit
because of the nature of Javascripts Number type. Avoiding binary
operations prevents this. Other solution would be using >>>0 (because it
internally converts the number to Uint32,
http://stackoverflow.com/questions/22335853/hack-to-convert-javascript-number-to-uint32)
like this.
((this._high << 16) | this._low) >>> 0
However binary operations are slower (tested in Node 5.x and Chrome 39).
Shifting 16 bits could be replaced by multiplying by 2^16=65536.
After that _high is spaced by 16 bits and _low is not gerater than those
16 bits so logical or could be simply replaced by simple addition.

toString() can leverage this and use native .toString() of JS Number
type for much faster result.